% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_covid19nytimes.R
\name{get_info_covid19nytimes}
\alias{get_info_covid19nytimes}
\title{Get information about the datasets provided by covid19nytimes}
\usage{
get_info_covid19nytimes()
}
\value{
a tibble of information about the datasets in this package
}
\description{
Returns information about the datasets in this package for covid19R harvesting
}
\examples{
\dontrun{

# get the dataset info from this package
get_info_covid19nytimes()
}

}
