% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh_covid19nytimes.R
\name{refresh_covid19nytimes_counties}
\alias{refresh_covid19nytimes_counties}
\title{US County Data from the NY Times}
\source{
New York Times, "We’re Sharing Coronavirus Case Data for Every U.S. County" \href{https://www.nytimes.com/article/coronavirus-county-data-us.html}{article} and {https://github.com/nytimes/covid-19-data}{data repository}

\href{https://github.com/Covid19R/documentation}{covid19R documentation}
}
\usage{
refresh_covid19nytimes_counties()
}
\value{
A tibble object
}
\description{
Update the county-level data from the New York Times on Covid-19
}
\details{
Pulls from the open source data at https://www.nytimes.com/article/coronavirus-county-data-us.html
including cummulative cases and deaths. Returns the data in the covid19R standard tidy format for easy use.
Location data is returned as \verb{County,State}.
See https://github.com/nytimes/covid-19-data for extensive methodology description and license.
}
\examples{
\dontrun{

#update the data
covid19nytimes_counties <- refresh_covid19nytimes_counties()


}
}
